backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
semantic-kernel==0.3.12.dev0
pydantic==1.10.12
